+++ libfreerdp/core/gateway/rpc_client.c
@@ -478,8 +478,7 @@ static int rpc_client_recv_fragment(rdpRpc* rpc, wStream* fragment)
 			if (Stream_Length(fragment) < StubOffset + StubLength)
 				goto fail;
 			Stream_SetPosition(fragment, StubOffset);
-			rpc_client_receive_pipe_write(rpc->client, Stream_ConstPointer(fragment),
-			                              (size_t)StubLength);
+			rpc_client_receive_pipe_write(rpc->client, Stream_ConstPointer(fragment), StubLength);
 			rpc->StubFragCount++;
 
 			if (response->alloc_hint == StubLength)
